New User Controls

+++ MP GUI VERSION/MPGuiVersion/MPGuiVersion/Base Classes/Item.cs
@@ -8,14 +8,14 @@ namespace MPGuiVersion
     {
         public string itemName;
         public int Quantity;
-        public int UnitPrice;
+        public float UnitPrice;
         public string Description;
 
         public string Product { get => itemName; set => itemName = value; }
 
         public int Ammount { get => Quantity; set => Quantity = value; }
 
-        public int Price { get => UnitPrice; set => UnitPrice = value; }
+        public float Price { get => UnitPrice; set => UnitPrice = value; }
 
         public string Summary { get => Description; set => Description = value; }
     }
